// apps/workers/src/types/tmdb-api.d.ts
export interface operations {
// ...
  'movie-details': {
    parameters: {
      query?: {
        /** @description comma separated list of endpoints within this namespace, 20 items max */
        append_to_response?: string;
        language?: string;
      };
      header?: never;
      path: {
        movie_id: number;
      };
      cookie?: never;
    };
    requestBody?: never;
    responses: {
      /** @description 200 */
      200: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          'application/json': {
            /**
             * @default false
             * @example false
             */
            adult: boolean;
            /** @example /2w4xG178RpB4MDAIfTkqAuSJzec.jpg */
            backdrop_path?: string;
            belongs_to_collection?: {
              /**
               * @default 0
               * @example 10
               */
              id: number;
              /** @example Star Wars Collection */
              name?: string;
              /** @example /pWVLFh4OuejTpUaDQbB1C4zoS2p.jpg */
              poster_path?: string;
              /** @example /iY2ujEY2m68OTTlPFTiHub9joHS.jpg */
              backdrop_path?: string;
            };
            /**
             * @default 0
             * @example 11000000
             */
            budget: number;
            genres?: {
              /**
               * @default 0
               * @example 12
               */
              id: number;
              /** @example Adventure */
              name?: string;
            }[];
            /** @example http://www.starwars.com/films/star-wars-episode-iv-a-new-hope */
            homepage?: string;
            /**
             * @default 0
             * @example 11
             */
            id: number;
            /** @example tt0076759 */
            imdb_id?: string;
            origin_country?: string[];
            /** @example en */
            original_language?: string;
            /** @example Star Wars */
            original_title?: string;
            /** @example Princess Leia is captured and held hostage by the evil Imperial forces in their effort to take over the galactic Empire. Venturesome Luke Skywalker and dashing captain Han Solo team together with the loveable robot duo R2-D2 and C-3PO to rescue the beautiful princess and restore peace and justice in the Empire. */
            overview?: string;
            /**
             * @default 0
             * @example 20.6912
             */
            popularity: number;
            /** @example /6FfCtAuVAW8XJjZ7eWeLibRLWTw.jpg */
            poster_path?: string;
            production_companies?: {
              /**
               * @default 0
               * @example 1
               */
              id: number;
              /** @example /tlVSws0RvvtPBwViUyOFAO0vcQS.png */
              logo_path?: string;
              /** @example Lucasfilm Ltd. */
              name?: string;
              /** @example US */
              origin_country?: string;
            }[];
            production_countries?: {
              /** @example US */
              iso_3166_1?: string;
              /** @example United States of America */
              name?: string;
            }[];
            /** @example 1977-05-25 */
            release_date?: string;
            /**
             * @default 0
             * @example 775398007
             */
            revenue: number;
            /**
             * @default 0
             * @example 121
             */
            runtime: number;
            spoken_languages?: {
              /** @example English */
              english_name?: string;
              /** @example en */
              iso_639_1?: string;
              /** @example English */
              name?: string;
            }[];
            /** @example Released */
            status?: string;
            /** @example A long time ago in a galaxy far, far away... */
            tagline?: string;
            /** @example Star Wars */
            title?: string;
            /**
             * @default false
             * @example false
             */
            video: boolean;
            /**
             * @default 0
             * @example 8.2
             */
            vote_average: number;
            /**
             * @default 0
             * @example 22061
             */
            vote_count: number;
          };
        };
      };
    };
  };
// ...
  'movie-videos': {
    parameters: {
      query?: {
        language?: string;
      };
      header?: never;
      path: {
        movie_id: number;
      };
      cookie?: never;
    };
    requestBody?: never;
    responses: {
      /** @description 200 */
      200: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          'application/json': {
            /**
             * @default 0
             * @example 550
             */
            id: number;
            results?: {
              /** @example en */
              iso_639_1?: string;
              /** @example US */
              iso_3166_1?: string;
              /** @example Fight Club (1999) Trailer - Starring Brad Pitt, Edward Norton, Helena Bonham Carter */
              name?: string;
              /** @example O-b2VfmmbyA */
              key?: string;
              /** @example YouTube */
              site?: string;
              /**
               * @default 0
               * @example 720
               */
              size: number;
              /** @example Trailer */
              type?: string;
              /**
               * @default true
               * @example false
               */
              official: boolean;
              /** @example 2016-03-05T02:03:14.000Z */
              published_at?: string;
              /** @example 639d5326be6d88007f170f44 */
              id?: string;
            }[];
          };
        };
      };
    };
  };
// ...
  'search-movie': {
    parameters: {
      query: {
        query: string;
        include_adult?: boolean;
        language?: string;
        primary_release_year?: string;
        page?: number;
        region?: string;
        year?: string;
      };
      header?: never;
      path?: never;
      cookie?: never;
    };
    requestBody?: never;
    responses: {
      /** @description 200 */
      200: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          'application/json': {
            /**
             * @default 0
             * @example 1
             */
            page: number;
            results?: {
              /**
               * @default true
               * @example false
               */
              adult: boolean;
              /** @example /hZkgoQYus5vegHoetLkCJzb17zJ.jpg */
              backdrop_path?: string;
              genre_ids?: number[];
              /**
               * @default 0
               * @example 550
               */
              id: number;
              /** @example en */
              original_language?: string;
              /** @example Fight Club */
              original_title?: string;
              /** @example A ticking-time-bomb insomniac and a slippery soap salesman channel primal male aggression into a shocking new form of therapy. Their concept catches on, with underground "fight clubs" forming in every town, until an eccentric gets in the way and ignites an out-of-control spiral toward oblivion. */
              overview?: string;
              /**
               * @default 0
               * @example 73.433
               */
              popularity: number;
              /** @example /pB8BM7pdSp6B6Ih7QZ4DrQ3PmJK.jpg */
              poster_path?: string;
              /** @example 1999-10-15 */
              release_date?: string;
              /** @example Fight Club */
              title?: string;
              /**
               * @default true
               * @example false
               */
              video: boolean;
              /**
               * @default 0
               * @example 8.433
               */
              vote_average: number;
              /**
               * @default 0
               * @example 26279
               */
              vote_count: number;
            }[];
            /**
             * @default 0
             * @example 2
             */
            total_pages: number;
            /**
             * @default 0
             * @example 39
             */
            total_results: number;
          };
        };
      };
    };
  };
// ...
}